requirements.txt
# Data collection
requests>=2.31.0
PyGithub>=2.1.1
yfinance>=0.2.31

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0

# Storage
pyyaml>=6.0.1
python-dateutil>=2.8.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
pytz>=2023.3

# Development (optional)
jupyter>=1.0.0
notebook>=7.0.0
